+++ src/dom-viewer/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0 (16 Dec 2022)
+
+* feat: support select given node
+* fix: remove  comma between attributes
+
 ## 1.0.2 (22 Nov 2022)
 
 * fix: hide empty text node